LayoutTests/media/media-source/media-source-real-seek-complete.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-seek-complete</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Real-MP4 equivalent of media-source-seek-complete.html
    // Tests that seeking completes after endOfStream.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        sourceBuffer = source.addSourceBuffer(MP4.VIDEO_TYPE);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        // One sample at PTS 10-11
        var media = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 10000, samples: [
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
            ]}]
        });
        sourceBuffer.appendBuffer(media);
        await waitFor(sourceBuffer, 'updateend');

        run('source.duration = 11');
        run('source.endOfStream()');
        run('video.currentTime = 10');
        await waitFor(video, 'seeked');

        endTest();
    }
    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
