LayoutTests/media/media-source/media-source-real-seek-complete-expected.txt


EVENT(sourceopen)
EVENT(updateend)
EVENT(updateend)
RUN(source.duration = 11)
RUN(source.endOfStream())
RUN(video.currentTime = 10)
EVENT(seeked)
END OF TEST
